+++ src/AngleSharp/Dom/ElementExtensions.cs
@@ -287,12 +287,7 @@ namespace AngleSharp.Dom
         /// <returns>True if the element is currently enabled, otherwise false.</returns>
         public static Boolean IsEnabled(this IElement element)
         {
-            if (element is IHtmlAnchorElement or IHtmlAreaElement or IHtmlLinkElement)
-            {
-                var href = element.GetAttribute(null, AttributeNames.Href);
-                return !String.IsNullOrEmpty(href);
-            }
-            else if (element is IHtmlButtonElement buttonElement)
+            if (element is IHtmlButtonElement buttonElement)
             {
                 return !buttonElement.IsDisabled;
             }
@@ -567,18 +562,11 @@ namespace AngleSharp.Dom
         {
             if (element is IHtmlAnchorElement anchor)
             {
-                var href = element.GetAttribute(null, AttributeNames.Href);
-                return !String.IsNullOrEmpty(href) && CheckVisited(anchor);
+                return element.HasAttribute(null, AttributeNames.Href) && CheckVisited(anchor);
             }
             else if (element is IHtmlAreaElement area)
             {
-                var href = element.GetAttribute(null, AttributeNames.Href);
-                return !String.IsNullOrEmpty(href) && CheckVisited(area);
-            }
-            else if (element is IHtmlLinkElement link)
-            {
-                var href = element.GetAttribute(null, AttributeNames.Href);
-                return !String.IsNullOrEmpty(href) && CheckVisited(link);
+                return element.HasAttribute(null, AttributeNames.Href) && CheckVisited(area);
             }
 
             return false;
@@ -617,18 +605,11 @@ namespace AngleSharp.Dom
         {
             if (element is IHtmlAnchorElement anchor)
             {
-                var href = element.GetAttribute(null, AttributeNames.Href);
-                return !String.IsNullOrEmpty(href) && !anchor.IsVisited();
+                return element.HasAttribute(null, AttributeNames.Href) && !anchor.IsVisited();
             }
             else if (element is IHtmlAreaElement area)
             {
-                var href = element.GetAttribute(null, AttributeNames.Href);
-                return !String.IsNullOrEmpty(href) && !area.IsVisited();
-            }
-            else if (element is IHtmlLinkElement link)
-            {
-                var href = element.GetAttribute(null, AttributeNames.Href);
-                return !String.IsNullOrEmpty(href) && !link.IsVisited();
+                return element.HasAttribute(null, AttributeNames.Href) && !area.IsVisited();
             }
 
             return false;