+++ src/AngleSharp/Html/Dom/Internal/HtmlButtonElement.cs
@@ -140,6 +140,24 @@ namespace AngleSharp.Html.Dom
             }
         }
 
+        /// <inheritdoc />
+        public override void DoFocus()
+        {
+            if (!IsDisabled)
+            {
+                IsFocused = true;
+            }
+        }
+
+        /// <inheritdoc />
+        public override void DoBlur()
+        {
+            if (IsFocused)
+            {
+                IsFocused = false;
+            }
+        }
+
         #endregion
 
         #region Helper