+++ src/AngleSharp/Html/Dom/Internal/HtmlSelectElement.cs
@@ -105,6 +105,24 @@ namespace AngleSharp.Html.Dom
             Options.Remove(index);
         }
 
+        /// <inheritdoc />
+        public override void DoFocus()
+        {
+            if (!IsDisabled)
+            {
+                IsFocused = true;
+            }
+        }
+
+        /// <inheritdoc />
+        public override void DoBlur()
+        {
+            if (IsFocused)
+            {
+                IsFocused = false;
+            }
+        }
+
         #endregion
 
         #region Internal Methods